added cutting scaffold and associations

+++ db/migrate/20201112205620_create_orders.rb
@@ -4,7 +4,7 @@ class CreateOrders < ActiveRecord::Migration[6.0]
       t.string :seller_name
       t.string :country_origin
       t.integer :units
-      t.integer :total_price
+      t.decimal :total_price
       t.date :date_order
       t.date :date_shipped
       t.boolean :multiple
